+++ src/_packages/utils.index.ts
@@ -1,7 +1,7 @@
 import { EMOJIS, EMOJIS_IN_CATEGORIES } from '../utils/emojis';
 import { isValidJsonString } from '../utils/isValidJsonString';
+import { extractAllBlocksFromMarkdown } from '../utils/markdown/extractAllBlocksFromMarkdown';
 import { extractAllListItemsFromMarkdown } from '../utils/markdown/extractAllListItemsFromMarkdown';
-import { extractBlocksFromMarkdown } from '../utils/markdown/extractBlocksFromMarkdown';
 import { extractOneBlockFromMarkdown } from '../utils/markdown/extractOneBlockFromMarkdown';
 import { removeContentComments } from '../utils/markdown/removeContentComments';
 import { removeMarkdownFormatting } from '../utils/markdown/removeMarkdownFormatting';
@@ -14,8 +14,8 @@ import { unwrapResult } from '../utils/unwrapResult';
 export {
     EMOJIS,
     EMOJIS_IN_CATEGORIES,
+    extractAllBlocksFromMarkdown,
     extractAllListItemsFromMarkdown,
-    extractBlocksFromMarkdown,
     extractOneBlockFromMarkdown,
     isValidJsonString,
     removeContentComments,